pyproject.toml
[project]
name = "glpi-dw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-postgres>=1.10.0",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "sqlfluff>=4.0.4",
]
